\name{ICsurv-package}
\alias{ICsurv-package}
\docType{package}
\title{
A statistical package for regression analysis of interval-censored data under the semiparametric proportional hazards (PH)  model
}
\description{
This package allows for semiparametric regression analysis of interval-censored data under the proportional hazards model (PH). Monotone splines are used to estimate the unknown nondecreasing 
cumulative baseline hazard function. Model fitting is completed through the implementation of an expectation maximization (EM) algorithm.
}
\details{
\tabular{ll}{
Package: \tab ICsurv\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-18\cr
}
}

\author{  Christopher S. McMahan  and Lianming Wang

Maintainers: Christopher S. McMahan <mcmaha2@clemson.edu>  and Lianming Wang <wangl@stat.sc.edu>
}